Create Posts component
